PCI: rearrange linux/pci.h
While working on adding platform device support to the NVMe driver,
I noticed that a lot of interfaces that are declared in linux/pci.h
are invisible when CONFIG_PCI is disabled.

Apparently the intention was originally that every such function would
come with an inline stub helper for !CONFIG_PCI, but these days most
of them don't. This requires adding lots of extra #ifdef checks in
driver code when the compiler's dead code elimination could normally
just drop the caller from the object file.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>