[SUBMITTED 20210927] clocksource: ti-dmtimer: select TIMER_OF when needed
When CONFIG_OF is enabled but CONFIG_TIMER_OF is not, building the
dm systimer driver produces a warning:

arm-linux-gnueabi-ld: warning: orphan section `__timer_of_table' from `drivers/clocksource/timer-ti-dm-systimer.o' being placed in section `__timer_of_table'

Address this by selecting the TIMER_OF symbol in this configuration.
If CONFIG_OF is disabled, the table entry is not created, so there
is no warning, and no reason to have the infrastructure.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>