bpf: avoid out-of-range warning
The check against RINGBUF_MAX_DATA_SZ is known to be unnecessary on
32-bit architectures, but clang points out the same is true on
some 64-bit architectures, depending on the page size:

kernel/bpf/ringbuf.c:162:24: error: result of comparison of constant 1099511431168 with expression of type '__u32' (aka 'unsigned int') is always false [-Werror,-Wtautological-constant-out-of-range-compare]
        if (attr->max_entries > RINGBUF_MAX_DATA_SZ)
            ~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~

While correct, this warning is not helpful, as the compiler still does
the right thing for all configurations. Add a cast to shut up the
warning.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>