[SUBMITTED 20211019] [RFC] x86: avoid -mtune=atom with clang
The clang optimization for atom results in a large number of objtool
warnings like

  drivers/video/hdmi.o: warning: objtool: hdmi_infoframe_check()+0x74: unreachable instruction
  lib/crypto/curve25519.o: warning: objtool: init_module()+0x12: unreachable instruction
  drivers/clk/bcm/clk-iproc-armpll.o: warning: objtool: iproc_armpll_setup()+0x10c: unreachable instruction
  drivers/video/fbdev/core/cfbfillrect.o: warning: objtool: cfb_fillrect()+0xa2: unreachable instruction
  drivers/clk/bcm/clk-iproc-pll.o: warning: objtool: iproc_pll_clk_setup()+0x36d: unreachable instruction
  drivers/clk/bcm/clk-iproc-asiu.o: warning: objtool: iproc_asiu_setup()+0x2cf: unreachable instruction

and in rare cases problems with the register allocator:

  arch/x86/crypto/curve25519-x86_64.c:610:3: error: inline assembly requires more registers than available

  error: ran out of registers during register allocation

All of those can be avoided by changing the -mtune= option for
clang, leaving the -march= option unchanged. It's really a bug
that should be fxied in llvm, but it's fairly clear that the
atom optimizations are not well exercised in llvm, so not
using them is probably the safe choice regardless.

Link: https://github.com/ClangBuiltLinux/linux/issues/1483
Signed-off-by: Arnd Bergmann <arnd@arndb.de>