seqlock: split out linux/seqlock_types.h
sequence counts and locks are used in a couple of central data structures
in the kernel (inode, signal_struct, zone, sock, ...), but including
linux/seqlock.h pulls in other headers that are not needed for most
users of the high-level structures.

Split out the type definitions into a separate header that only
has minimal dependencies and can be included when defining a
structure based on sequence counts.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>