treewide: replace mmap_pgoff() with mmap2()
32-bit architectures generally implement sys_mmap2() as their primary
mmap() implementation, but common code only provides sys_mmap_pgoff(),
which has slightly different behavior when dealing with page sizes
other than 4096 bytes.

On ia64, m68k, arc and hexagon, the use of mmap_pgoff() appears to be
irreversible, as they do support multiple page sizes and user space
now relies on passing the argument in actual pages rather than
fixed-size units. m68k also has a misleading comment about SUN3
that needs to be rewritten to reflect that this is not broken but
required by libc now.

powerpc and riscv are additional exceptions here, as they have
checks for the 'prot' argument that need to be kept in place.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>