ARM: mmp: select specific CPU implementation
The behavior of the MMP platform code depends on whether the
CPU_PXA168/CPU_PXA910/CPU_MMP2 symbols are enabled or not.

I believe the intention here was that these can be left disabled for
a pure DT-only build, but it's not clear if that actually works. At
the minimum, the cpu_is_pxa168() and cpu_is_pxa910() checks behave
differently, which causes changes in the power management code.

For the moment, make the behavior depend on whether CONFIG_ATAGS is set
or not, to make it easier to bisect the removal of the old code later.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>