ARM: disallow big-endian armv6
ARMv5 is only allowed to be big-endian (BE32) on platforms that
are known to support this, while ARMv6 and ARMv7 can be big-endian
(BE8) on anything that is not known to be broken, as BE8 mode
is much easier to support.

When combinding ARMv6 and ARMv5 in a single kernel, this cannot
be big-endian at all because of the BE8/BE32 mismatch, so it's
easier to just disallow BE8 mode for ARMv6 entirely. There
are no known users of BE8 ARMv6, but ixp4xx is used in BE32
mode.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>