[VARIANT A] arm64: drop page size requirement on CONFIG_COMPAT
A change in binutils-2.40 has led to a regression on Debian Bookworm
arm32 userspace, which can no longer be run on kernels with 16KB page
size or larger including the Apple M1 series. This includes running with
qemu-user as there is a check to ensure that any library was built with
a sufficient alignment.

Link: https://sourceware.org/bugzilla/show_bug.cgi?id=30033
Signed-off-by: Arnd Bergmann <arnd@arndb.de>