ARM: skip 'setend' for ARMv7-M
Big-endian (BE8) kernels always switch into big-endian mode on the
initial entry in the decompressor or when waking up from sleep.

On ARMv7-M, this instruction is not available, which can be seen
from a build failure with clang:

arch/arm/boot/compressed/head.S:224:2: error: invalid instruction
 setend be @ go BE8 if compiled for BE8
arch/arm/kernel/sleep.S:155:1: error: invalid instruction
setend be @ ensure we are in BE mode

This corner case was not covered when the initial ARMv7-M support
and the BE8 support were added, so it is likely that neither sleep
mode nor compressed kernels ever worked in this configuration. The
entry point for the uncompressed kernel however does not have this
problem since the code in head{-nommu,}.S is not shared between
ARMv7-R and ARMv7-M.

This could be solved by either disallowing the configuration through
Kconfig or by fixing the code. It is unlikely that anyone cares
either way, so just do the simple code fix. This is obviously
untested because it would only work on a chip that is hardwired
into big-endian mode.

Fixes: c20611df13c3 ("ARM: 8327/1: zImage: add support for ARMv7-M")
Fixes: 97bcb0fea590 ("ARM: set BE8 if LE in head code")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>