radix-tree: move declarations to header
The xarray.c file contains the only call to radix_tree_node_rcu_free(),
and it comes with its own extern declaration for it. This means the
function definition causes a missing-prototype warning:

lib/radix-tree.c:288:6: error: no previous prototype for 'radix_tree_node_rcu_free' [-Werror=missing-prototypes]

Instead, move the declaration for this function to a new header that
can be included by both, and do the same for the radix_tree_node_cachep
variable that has the same underlying problem but does not cause
a warning with gcc.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>