x86: head: add dummy prototype for mk_early_pgtbl_32
'make W=1' warns about a function without a prototype in the x86-32 head code:

arch/x86/kernel/head32.c:72:13: error: no previous prototype for 'mk_early_pgtbl_32' [-Werror=missing-prototypes]

This is called from assembler code, so it does not actually need a prototype.
I could not find an appropriate header for it, so just declare it in front
of the definition to shut up th warning.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>