ARM: spear: remove unused spear13xx_l2x0_init() function
The L2 cache is defined in the dts file and will be initialized implicitly
through l2x0_of_init(), while the custom function was apparently never
called in mainline kernels and only remains as a leftover from early
upstreaming.

To actually use the pl310 cache controller, the CACHE_L2X0 symbol needs
to be enabled, and hte machine descriptor must set the l2c_aux_mask. Set
this to the default value, which assumes that the firmware has already
set it up correctly and the DT information is correct.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>