ARM: l2x0: hide CONFIG_CACHE_L2X0 behind CONFIG_EXPERT
There is little reason to disable the L2 cache controller support
on machines that have this hardware. To ensure this is always
enabled if any of the machines have it, change the dependencies
to only show the Kconfig option if CONFIG_EXPERT is set in
addition to CONFIG_MIGHT_HAVE_CACHE_L2X0.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>