[SUBMITTED 20230811] [RFC] Makefile.extrawarn: turn on missing-prototypes again
Over the years we went from > 1000 of warnings to under 100 earlier
this year, and I sent patches to address all the ones that I saw with
compile testing randcom configs on arm64, arm and x86 kernels. This is a
really useful warning, as it catches real bugs when there are mismatched
prototypes. In particular with kernel control flow integrity enabled,
those are no longer allowed.

There are still a few remaining warnings, and a couple of architectures
that have a lot of warnings in architecture specific code, so I'm
still unsure about whether we just enable this globally and have
architecture maintainers take care of the output, or we make the option
conditional based on the architecture.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>