syscalls: always include syscall prototypes
The sys_ni_posix_timers() definition causes a warning when
the declaration is missing, so this needs to be added
along with the normal syscalls, and the #ifdef around it
removed.

kernel/time/posix-stubs.c:26:17: error: no previous prototype for 'sys_ni_posix_timers' [-Werror=missing-prototypes]

As far as I can tell, the #ifdef was originally required when the sys_*
function namespace was overloaded on architectures using custom wrappers,
but that is no longer the case, so now we can actually make use of them
for the purpose of checking the interface and avoiding that warning.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>