irq: mxs: fix merge conflict for icoll_handle_irq() prototype
Two cleanup patches clashed, causing a build failure:

drivers/irqchip/irq-mxs.c:12:10: fatal error: 'linux/irqchip/mxs.h' file not found

I added the #include statement here to avoid a harmless warning,
while Linus removed the header file after it became obsolete.

The state we want in the next kernel is to make the function static
and not need the prototype.

Fixes: 0953fb263714e ("irq: remove handle_domain_{irq,nmi}()")
Fixes: 5bb578a0c1b86 ("ARM: 9298/1: Drop custom mdesc->handle_irq()")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
Unfortunately the two patches are merged into different trees, and
applying this patch to either of them would cause a new regression for
that individual tree, not sure how to best handle this.

The easiest way for now is probably to just revert my 0953fb263714e
patch in the irqchip tree for the moment, bringing back the original
warning until both are merged.