[SUBMITTED 20230705] btrfs: fix 64-bit division link failure
Some of the recent refactoring of the statfs code apparently
brought back a link failure on older gcc versions that I had
fixed before:

arm-linux-gnueabi-ld: fs/btrfs/super.o: in function `btrfs_statfs':
super.c:(.text+0xec40): undefined reference to `__aeabi_uldivmod'

I think what happened is that gcc is free to not inline a function
despite the 'inline' annotation, and when this happens it can end
up partially inlining the div_u64() helper in a way that breaks the
__builtin_constant_p() based optimization.

I only see this behavior for gcc-9, but it's possible that the same
thing happens in later versions as well when the code changes again.

Change this to __always_inline to prevent it from happening again,
and add a comment about this.

Fixes: 7e17916b35797 ("btrfs: avoid link error with CONFIG_NO_AUTO_INLINE")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>